# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Arkoo CRM API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Lead {
  id: string;
  name: string;
  email: string;
  phone: string;
  /** @nullable */
  source?: string | null;
  /** Current lead status (e.g. New, Contacted, Qualified, Converted, Lost) */
  status: string;
  /** AI-generated score from 0 to 100 */
  ai_score: number;
  /** AI-generated label (Hot, Warm, Cold) */
  ai_label: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  assigned_to?: string | null;
  /** @nullable */
  course_interest?: string | null;
  created_at: string;
  /** @nullable */
  updated_at?: string | null;
}

export interface LeadUpdate {
  status?: string;
  notes?: string;
  assigned_to?: string;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface LeadStats {
  total: number;
  hot: number;
  warm: number;
  cold: number;
  avg_score?: number;
  by_status: StatusCount[];
}

export type ListLeadsParams = {
/**
 * Filter by lead status
 */
status?: string;
/**
 * Filter by AI label (Hot, Warm, Cold)
 */
label?: string;
/**
 * Search by name, email or phone
 */
search?: string;
};