# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Arkoo CRM API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: leads
    description: Lead management operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List all leads
      description: Returns all leads with AI score and label
      parameters:
        - name: status
          in: query
          schema:
            type: string
          description: Filter by lead status
        - name: label
          in: query
          schema:
            type: string
          description: Filter by AI label (Hot, Warm, Cold)
        - name: search
          in: query
          schema:
            type: string
          description: Search by name, email or phone
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"

  /leads/stats:
    get:
      operationId: getLeadsStats
      tags: [leads]
      summary: Get lead statistics
      description: Returns aggregated counts by label and status
      responses:
        "200":
          description: Lead statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadStats"

  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lead details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Lead not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update lead status or notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Lead not found

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Lead:
      type: object
      required:
        - id
        - name
        - email
        - phone
        - status
        - ai_score
        - ai_label
        - created_at
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        source:
          type: ["string", "null"]
        status:
          type: string
          description: Current lead status (e.g. New, Contacted, Qualified, Converted, Lost)
        ai_score:
          type: number
          description: AI-generated score from 0 to 100
        ai_label:
          type: string
          description: AI-generated label (Hot, Warm, Cold)
        notes:
          type: ["string", "null"]
        assigned_to:
          type: ["string", "null"]
        course_interest:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: ["string", "null"]

    LeadUpdate:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
        assigned_to:
          type: string

    LeadStats:
      type: object
      required:
        - total
        - hot
        - warm
        - cold
        - by_status
      properties:
        total:
          type: integer
        hot:
          type: integer
        warm:
          type: integer
        cold:
          type: integer
        avg_score:
          type: number
        by_status:
          type: array
          items:
            $ref: "#/components/schemas/StatusCount"

    StatusCount:
      type: object
      required:
        - status
        - count
      properties:
        status:
          type: string
        count:
          type: integer